task list app

+++ Projects/todo-list/src/app/components/footer/footer.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { inputTask, makeTodo, deleteAllItens } from '../utils/todoFactoryInplementation';
+import { inputTask, listCheckedTask, makeTodo, deleteAllItens, sendToDBTodo, setTodos } from '../../services/todoFactoryInplementation';
 
 @Component({
   selector: 'app-footer',
@@ -10,17 +10,24 @@ import { inputTask, makeTodo, deleteAllItens } from '../utils/todoFactoryInpleme
 export class FooterComponent implements OnInit {
   
   public inputTask :any = null
+  public listCheckedTask :any = null
+  public inputTaskElement :HTMLInputElement | null = null
   
   ngDoCheck(): void {
     this.inputTask = inputTask
+    this.listCheckedTask = listCheckedTask
   }
 
-  
-  public inputTaskElement :HTMLInputElement | null = null
-
   ngOnInit(): void {
-    
     this.inputTaskElement = document.querySelector('#inputTask') as HTMLInputElement | null 
+    
+    try {
+      setTodos()
+    } catch (error) {
+      console.log('hmm');
+    }
+
+
     this.inputTaskElement?.addEventListener('keydown', (event)=> {
       
       if(event.key === 'Enter' && this.inputTaskElement?.value != undefined ){
@@ -28,15 +35,21 @@ export class FooterComponent implements OnInit {
         if(this.inputTaskElement?.value === '') return
         
         makeTodo(this.inputTaskElement?.value as string)
+        sendToDBTodo()
+
         this.inputTaskElement.value = ''
       };
       
     })
 
-  }  
+  }
 
   public deleteAll() :void{
+
+    if ( !window.confirm('deseja mesmo deletar tudo?') ) return
+     
     deleteAllItens();
+    sessionStorage.setItem("todoArray", '' )
   }
 
 }
